Multi-Env-ASG-Deployment/3-tier-user-management-app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dotenv==1.1.1
python-dotenv